Add an env reader that checks for secret files (#424)
* Add an env reader that checks for secret files

Based on the work by @avilaton in https://github.com/sloria/environs/pull/295

* Add example file

* When the env var is expected to be a file path, it must be available

We raise ValueError, similar to what happens when a referenced .env file can't be read

* More tests, fix incorrect super() behaviour

When `self` is omitted, weird things happen, so don't declare it a static method.

* Add types

* Test file-aware variable expansion

Nothing really out of the ordinary, but this is an actual use case of course.

* Add support for specifying a custom suffix for file env vars

* Add README explanation on FileAwareEnv

* Update README.md

* Update README.md

* Update src/environs/__init__.py

* Update src/environs/__init__.py

* fixups

* Update changelog

---------

Co-authored-by: Steven Loria <sloria1@pm.me>
Co-authored-by: Steven Loria <sloria1@gmail.com>